Get CHIRPS precipitation data

Get daily precipitation data from the "Climate Hazards Group InfraRed Precipitation with Station Data" via ClimateSERV API client. ClimateSERV works with geojson of type 'Polygon'. The input object is then transformed into polygons with a small buffer area around the point.

get_chirps(object, dates, operation = 5, ...)

# S3 method for default
get_chirps(object, dates, operation = 5, ...)

# S3 method for sf
get_chirps(object, dates, operation = 5, as.sf = FALSE, ...)

# S3 method for geojson
get_chirps(object, dates, operation = 5, as.geojson = FALSE, ...)

Arguments

object
input, an object of class data.frame,geojson or sf

dates
a character of start and end dates in that order in the format "YYYY-MM-DD"

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

...
further arguments passed to sf methods. See details

as.sf
logical, returns an object of class [sf]{sf} for S3 method of chirps class sf

as.geojson
logical, returns an object of class geojson for S3 method of chirps class geojson

Value

A data frame of CHIRPS data:

id

the index for the rows in object

dates

the dates from which CHIRPS was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

chirps

the CHIRPS value in mm

Details

operation: supported operations are:

operation		value

max	=	0

min	=	1

median	=	2

sum	=	4

average	=	5 (default value)

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

References

Funk C. et al. (2015). Scientific Data, 2, 150066.
https://doi.org/10.1038/sdata.2015.66

ClimateSERV https://climateserv.servirglobal.net

Examples

# \donttest{ 
library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15", "2017-12-31")

get_chirps(lonlat, dates)

#> Getting your request...

#> # A tibble: 3 x 5
#>   id      lon   lat date       chirps
#>   <chr> <dbl> <dbl> <date>      <dbl>
#> 1 1     -55.0 -2.81 2017-12-15    0  
#> 2 2     -55.0 -2.88 2017-12-16    0  
#> 3 3     -55.1 -3.53 2017-12-17   13.7

############################################

# S3 method for objects of class 'sf'
library("sf")

#> Linking to GEOS 3.7.2, GDAL 2.4.2, PROJ 5.2.0

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))

dates <- c("2017-12-15", "2017-12-31")

get_chirps(lonlat, dates)

#> Warning: dist is assumed to be in decimal degrees (arc_degrees)

#> Getting your request...

#> # A tibble: 3 x 5
#>   id      lon   lat date       chirps
#>   <chr> <dbl> <dbl> <date>      <dbl>
#> 1 1     -55.0 -2.81 2017-12-15    0  
#> 2 2     -55.0 -2.88 2017-12-16    0  
#> 3 3     -55.1 -3.53 2017-12-17   13.7

# as.sf = TRUE returns an object of class 'sf'
get_chirps(lonlat, dates, as.sf = TRUE)

#> Warning: dist is assumed to be in decimal degrees (arc_degrees)

#> Getting your request...

#> Simple feature collection with 3 features and 17 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -55.0714 ymin: -3.5279 xmax: -54.9857 ymax: -2.8094
#> epsg (SRID):    NA
#> proj4string:    NA
#>   d_17515 d_17516  d_17517  d_17518 d_17519 d_17520 d_17521 d_17522 d_17523
#> 1       0       0 29.59562 14.79781       0       0       0       0       0
#> 2       0       0 13.68630 22.12671       0       0       0       0       0
#> 3       0       0 11.06336 13.68630       0       0       0       0       0
#>    d_17524  d_17525  d_17526  d_17527   d_17528 d_17529   d_17530   d_17531
#> 1 7.373443 0.000000 0.000000 37.68985 18.844923       0  9.422462  9.422462
#> 2 0.000000 0.000000 9.385599 37.54240 18.771198       0  9.385599  9.385599
#> 3 0.000000 6.649946 0.000000 22.16965  5.542412       0 11.084825 11.084825
#>                   geometry
#> 1 POINT (-55.0281 -2.8094)
#> 2 POINT (-54.9857 -2.8756)
#> 3 POINT (-55.0714 -3.5279)

############################################

# S3 method for objects of class 'geojson'
library("geojsonsf")

#> Error in library("geojsonsf"): there is no package called ‘geojsonsf’

library("sf")

tapajos <- chirps:::tapajos

object <- chirps:::.sf_to_geojson(tapajos[1:2,])

#> Warning: st_buffer does not correctly buffer longitude/latitude data

#> dist is assumed to be in decimal degrees (arc_degrees).

object <- unlist(object)

class(object) <- c("geojson", "json", class(object))

dates <- c("2018-01-01","2018-01-20")

get_chirps(object, dates)

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Getting your request...

#> # A tibble: 40 x 5
#>    id      lon   lat date       chirps
#>    <chr> <dbl> <dbl> <date>      <dbl>
#>  1 1     -55.0 -3.36 2018-01-01    0  
#>  2 1     -55.0 -3.36 2018-01-02    0  
#>  3 1     -55.0 -3.36 2018-01-03    0  
#>  4 1     -55.0 -3.36 2018-01-04    0  
#>  5 1     -55.0 -3.36 2018-01-05    0  
#>  6 1     -55.0 -3.36 2018-01-06    0  
#>  7 1     -55.0 -3.36 2018-01-07    0  
#>  8 1     -55.0 -3.36 2018-01-08   16.7
#>  9 1     -55.0 -3.36 2018-01-09    0  
#> 10 1     -55.0 -3.36 2018-01-10    0  
#> # … with 30 more rows

# as.geojson = TRUE returns an object of class 'geojson'
get_chirps(object, dates, as.geojson = TRUE)

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Getting your request...

#> {"type":"FeatureCollection","name":"chirps","features":[{"type":"Feature","properties":{"1":{"id":"1","date":"2018-01-01","value":"0"},"10":{"id":"1","date":"2018-01-10","value":"0"},"11":{"id":"1","date":"2018-01-11","value":"0"},"12":{"id":"1","date":"2018-01-12","value":"8.16110038757324"},"13":{"id":"1","date":"2018-01-13","value":"8.16110038757324"},"14":{"id":"1","date":"2018-01-14","value":"16.3222007751465"},"15":{"id":"1","date":"2018-01-15","value":"0"},"16":{"id":"1","date":"2018-01-16","value":"5.37311792373657"},"17":{"id":"1","date":"2018-01-17","value":"5.37311792373657"},"18":{"id":"1","date":"2018-01-18","value":"16.1193542480469"},"19":{"id":"1","date":"2018-01-19","value":"0"},"2":{"id":"1","date":"2018-01-02","value":"0"},"20":{"id":"1","date":"2018-01-20","value":"0"},"3":{"id":"1","date":"2018-01-03","value":"0"},"4":{"id":"1","date":"2018-01-04","value":"0"},"5":{"id":"1","date":"2018-01-05","value":"0"},"6":{"id":"1","date":"2018-01-06","value":"0"},"7":{"id":"1","date":"2018-01-07","value":"0"},"8":{"id":"1","date":"2018-01-08","value":"16.741189956665"},"9":{"id":"1","date":"2018-01-09","value":"0"}},"geometry":{"type":["Polygon"],"coordinates":["-55.0291869966114","-55.0291899255435","-55.0291969966114","-55.0292040676792","-55.0292069966114","-55.0292040676792","-55.0291969966114","-55.0291899255435","-55.0291869966114","-3.36021942199265","-3.36022649306046","-3.36022942199265","-3.36022649306046","-3.36021942199265","-3.36021235092484","-3.36020942199265","-3.36021235092484","-3.36021942199265"]}}]} {"type":"FeatureCollection","name":"chirps","features":[{"type":"Feature","properties":{"21":{"id":"2","date":"2018-01-01","value":"0"},"22":{"id":"2","date":"2018-01-02","value":"0"},"23":{"id":"2","date":"2018-01-03","value":"0"},"24":{"id":"2","date":"2018-01-04","value":"0"},"25":{"id":"2","date":"2018-01-05","value":"0"},"26":{"id":"2","date":"2018-01-06","value":"0"},"27":{"id":"2","date":"2018-01-07","value":"0"},"28":{"id":"2","date":"2018-01-08","value":"18.883243560791"},"29":{"id":"2","date":"2018-01-09","value":"0"},"30":{"id":"2","date":"2018-01-10","value":"0"},"31":{"id":"2","date":"2018-01-11","value":"0"},"32":{"id":"2","date":"2018-01-12","value":"7.42914581298828"},"33":{"id":"2","date":"2018-01-13","value":"7.42914581298828"},"34":{"id":"2","date":"2018-01-14","value":"22.2874336242676"},"35":{"id":"2","date":"2018-01-15","value":"7.42914581298828"},"36":{"id":"2","date":"2018-01-16","value":"5.55930471420288"},"37":{"id":"2","date":"2018-01-17","value":"5.55930471420288"},"38":{"id":"2","date":"2018-01-18","value":"16.6779136657715"},"39":{"id":"2","date":"2018-01-19","value":"0"},"40":{"id":"2","date":"2018-01-20","value":"0"}},"geometry":{"type":["Polygon"],"coordinates":["-55.0371837306169","-55.0371866595491","-55.0371937306169","-55.0372008016847","-55.0372037306169","-55.0372008016847","-55.0371937306169","-55.0371866595491","-55.0371837306169","-3.23339364149234","-3.23340071256015","-3.23340364149234","-3.23340071256015","-3.23339364149234","-3.23338657042453","-3.23338364149234","-3.23338657042453","-3.23339364149234"]}}]} 

# }