Get evaporative stress index (ESI) data

Get evaporative stress index (ESI) from SERVIR Global via ClimateSERV API Client. ESI is available every four (or twelve) weeks from 2001 to present. The dataset may contain cloudy data which is returned as NAs. ClimateSERV works with geojson of type 'Polygon'. The input object is then transformed into polygons with a small buffer area around the point.

get_esi(object, dates, operation = 5, period = 1, ...)

# S3 method for default
get_esi(object, dates, operation = 5, period = 1, ...)

# S3 method for sf
get_esi(object, dates, operation = 5, period = 1, as.sf = FALSE, ...)

# S3 method for geojson
get_esi(object, dates, operation = 5, period = 1, as.geojson = FALSE, ...)

Arguments

object
input, an object of class data.frame,geojson or sf

dates
a character of start and end dates in that order in the format "YYYY-MM-DD"

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

period
an integer value for the period of ESI data, four weeks period = 1, twelve weeks = 2

...
further arguments passed to sf methods. See details

as.sf
logical, returns an object of class [sf]{sf} for S3 method of chirps class sf

as.geojson
logical, returns an object of class geojson for S3 method of chirps class geojson

Value

A data frame of ESI data:

id

the index for the rows in object

dates

the dates from which ESI was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

esi

the ESI value

Details

operation: supported operations are:

operation		value

max	=	0

min	=	1

median	=	2

sum	=	4

average	=	5 (default value)

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

References

ClimateSERV https://climateserv.servirglobal.net

Evaporative Stress Index http://catalogue.servirglobal.net/Product?product_id=198

Examples

# \donttest{
library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15","2018-06-20")

# by default the function set a very small buffer around the points
# which can return NAs due to cloudiness in ESI data

dat <- get_esi(lonlat, dates = dates)

#> Getting your request...

dat

#> # A tibble: 3 x 5
#>   id      lon   lat date          esi
#>   <chr> <dbl> <dbl> <date>      <dbl>
#> 1 1     -55.0 -2.81 2017-12-11 -0.844
#> 2 2     -55.0 -2.88 2018-01-01 -1.51 
#> 3 3     -55.1 -3.53 2018-01-09 -0.536

# the argument dist passed through sf increase the buffer area

dat <- get_esi(lonlat, dates = dates, dist = 0.1)

#> Getting your request...

dat

#> # A tibble: 3 x 5
#>   id      lon   lat date          esi
#>   <chr> <dbl> <dbl> <date>      <dbl>
#> 1 1     -55.0 -2.81 2017-12-11 -1.98 
#> 2 2     -55.0 -2.88 2018-01-01 -1.49 
#> 3 3     -55.1 -3.53 2018-01-09 -0.219

############################################

# S3 method for objects of class 'sf'
library("sf")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))

dates <- c("2017-12-15", "2017-12-31")

get_esi(lonlat, dates)

#> Getting your request...

#> # A tibble: 1 x 5
#>   id      lon   lat date          esi
#>   <chr> <dbl> <dbl> <date>      <dbl>
#> 1 1     -55.0 -2.81 2017-12-11 -0.844

# as.sf = TRUE returns an object of class 'sf'
get_esi(lonlat, dates, as.sf = TRUE)

#> Getting your request...

#> Simple feature collection with 3 features and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -55.0714 ymin: -3.5279 xmax: -54.9857 ymax: -2.8094
#> epsg (SRID):    NA
#> proj4string:    NA
#>      d_17511                 geometry
#> 1 -0.8444786 POINT (-55.0281 -2.8094)
#> 2 -0.8444786 POINT (-54.9857 -2.8756)
#> 3 -0.8444786 POINT (-55.0714 -3.5279)

############################################

# S3 method for objects of class 'geojson'
library("geojsonsf")

#> Error in library("geojsonsf"): there is no package called ‘geojsonsf’

library("sf")

tapajos <- chirps:::tapajos

object <- chirps:::.sf_to_geojson(tapajos[1:2,])

#> Warning: st_buffer does not correctly buffer longitude/latitude data

#> dist is assumed to be in decimal degrees (arc_degrees).

object <- unlist(object)

class(object) <- c("geojson", "json", class(object))

dates <- c("2018-01-01","2018-01-20")

get_esi(object, dates)

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Getting your request...

#> # A tibble: 6 x 5
#>   id      lon   lat date           esi
#>   <chr> <dbl> <dbl> <date>       <dbl>
#> 1 1     -55.0 -3.36 2018-01-01 -0.533 
#> 2 1     -55.0 -3.36 2018-01-09 -0.291 
#> 3 1     -55.0 -3.36 2018-01-17 -0.0541
#> 4 2     -55.0 -3.23 2018-01-01 -0.142 
#> 5 2     -55.0 -3.23 2018-01-09 -0.247 
#> 6 2     -55.0 -3.23 2018-01-17 -0.149 

# as.geojson = TRUE returns an object of class 'geojson'
get_esi(object, dates, as.geojson = TRUE)

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Warning: st_centroid does not give correct centroids for longitude/latitude data

#> Getting your request...

#> {"type":"FeatureCollection","name":"esi","features":[{"type":"Feature","properties":{"1":{"id":"1","date":"2018-01-01","value":"-0.532923996448517"},"2":{"id":"1","date":"2018-01-09","value":"-0.290530353784561"},"3":{"id":"1","date":"2018-01-17","value":"-0.0540637709200382"}},"geometry":{"type":["Polygon"],"coordinates":["-55.0291869966114","-55.0291899255435","-55.0291969966114","-55.0292040676792","-55.0292069966114","-55.0292040676792","-55.0291969966114","-55.0291899255435","-55.0291869966114","-3.36021942199265","-3.36022649306046","-3.36022942199265","-3.36022649306046","-3.36021942199265","-3.36021235092484","-3.36020942199265","-3.36021235092484","-3.36021942199265"]}}]} {"type":"FeatureCollection","name":"esi","features":[{"type":"Feature","properties":{"4":{"id":"2","date":"2018-01-01","value":"-0.142488121986389"},"5":{"id":"2","date":"2018-01-09","value":"-0.247443631291389"},"6":{"id":"2","date":"2018-01-17","value":"-0.148515313863754"}},"geometry":{"type":["Polygon"],"coordinates":["-55.0371837306169","-55.0371866595491","-55.0371937306169","-55.0372008016847","-55.0372037306169","-55.0372008016847","-55.0371937306169","-55.0371866595491","-55.0371837306169","-3.23339364149234","-3.23340071256015","-3.23340364149234","-3.23340071256015","-3.23339364149234","-3.23338657042453","-3.23338364149234","-3.23338657042453","-3.23339364149234"]}}]} 

# }