Get Integrated Multisatellite Retrievals for GPM (IMERG) data

The IMERG dataset provides near-real time global observations of rainfall at 10km resolution, which can be used to estimate total rainfall accumulation from storm systems and quantify the intensity of rainfall and flood impacts from tropical cyclones and other storm systems. IMERG is a daily precipitation dataset available from 2015 to present within the latitudes 70 and -70.

get_imerg(object, dates, operation = 5, ...)

# S3 method for default
get_imerg(object, dates, operation = 5, ...)

# S3 method for sf
get_imerg(object, dates, operation = 5, as.sf = FALSE, ...)

# S3 method for geojson
get_imerg(object, dates, operation = 5, as.geojson = FALSE, ...)

Arguments

object
input, an object of class data.frame (or any other object that can be coerced to data.frame), SpatVector, SpatRaster, sf or geojson

dates
a character of start and end dates in that order in the format "YYYY-MM-DD"

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

...
further arguments passed to terra or sf methods See details

as.sf
logical, returns an object of class sf

as.geojson
logical, returns an object of class geojson

Value

A data frame of imerg data:

id

the index for the rows in object

dates

the dates from which imerg was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

imerg

the IMERG value

Details

operation: supported operations are:

operation		value

max	=	0

min	=	1

median	=	2

sum	=	4

average	=	5 (default value)

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

Examples

# \donttest{
lonlat <- data.frame(lon = c(-55.0281,-54.9857),
                     lat = c(-2.8094, -2.8756))

dates <- c("2017-12-15", "2017-12-31")

dt <- get_imerg(lonlat, dates)

#> Fetching data from ClimateSERV server 

#> Getting your request...

dt

#>        id    lon   lat       date   imerg
#>     <int>  <dbl> <dbl>     <date>   <dbl>
#> 1:      1 -55.03 -2.81 2017-12-15   48.00
#> 2:      1 -55.03 -2.81 2017-12-16    0.00
#> 3:      1 -55.03 -2.81 2017-12-17  273.00
#> 4:      1 -55.03 -2.81 2017-12-18   93.00
#> 5:      1 -55.03 -2.81 2017-12-19  208.00
#> ---                                      
#> 30:     2 -54.99 -2.88 2017-12-27 1274.00
#> 31:     2 -54.99 -2.88 2017-12-28  208.00
#> 32:     2 -54.99 -2.88 2017-12-29   13.00
#> 33:     2 -54.99 -2.88 2017-12-30  174.00
#> 34:     2 -54.99 -2.88 2017-12-31  100.00

# }