API Client for CHIRPS

API Client for the Climate Hazards Group InfraRed Precipitation with Station Data 'CHIRPS'. The 'CHIRPS' data is a 35+ year quasi-global rainfall data set, which incorporates 0.05 arc-degrees resolution satellite imagery, and in-situ station data to create gridded rainfall time series for trend analysis and seasonal drought monitoring. For more details on 'CHIRPS' data please visit its official home page <https://www.chc.ucsb.edu/data/chirps>. Requests from large time series (> 10 years) and large geographic coverage (global scale) may take several minutes.

Note

While chirps does not redistribute the data or provide it in any way, we encourage users to cite Funk et al. (2015) when using the CHIRPS data.

Funk C., et al. (2015). Scientific Data, 2, 150066. https://doi.org/10.1038/sdata.2015.66

See also

Author

Kauê de Sousa and Adam H. Sparks