Get CHIRPS precipitation data

Get daily precipitation data from the "Climate Hazards Group InfraRed Precipitation with Station Data" via ClimateSERV API client. ClimateSERV works with geojson of type 'Polygon'. The input object is then transformed into polygons with a small buffer area around the point.

get_chirps(object, dates, operation = 5, ...)

# S3 method for default
get_chirps(object, dates, operation = 5, ...)

# S3 method for sf
get_chirps(object, dates, operation = 5, as.sf = FALSE, ...)

Arguments

object
input, an object of class data.frame or sf

dates
a character of start and end dates in that order in the format "YYYY-MM-DD"

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

...
further arguments passed to sf methods. See details

as.sf
logical, returns an object of class [sf]{sf} for S3 method of chirps class sf

Value

A data frame of CHIRPS data:

id

the index for the rows in object

dates

the dates from which CHIRPS was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

chirps

the CHIRPS value in mm

Details

operation: supported operations are:

operation

value

max	=	0

min	=

1		median

=	2

sum	=	4

average	=

5 (	default value	)

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

References

Funk C. et al. (2015). Scientific Data, 2, 150066.
https://doi.org/10.1038/sdata.2015.66

ClimateSERV https://climateserv.servirglobal.net

Examples

# Three points in the Tapajos National Forest, Brazil
library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15","2017-12-31")

get_chirps(lonlat, dates)

#> Getting your request...

#> # A tibble: 51 x 5
#>    id      lon   lat date       chirps
#>    <chr> <dbl> <dbl> <date>      <dbl>
#>  1 1     -55.0 -2.81 2017-12-15   0   
#>  2 1     -55.0 -2.81 2017-12-16   0   
#>  3 1     -55.0 -2.81 2017-12-17  13.7 
#>  4 1     -55.0 -2.81 2017-12-18  13.7 
#>  5 1     -55.0 -2.81 2017-12-19   0   
#>  6 1     -55.0 -2.81 2017-12-20   0   
#>  7 1     -55.0 -2.81 2017-12-21   0   
#>  8 1     -55.0 -2.81 2017-12-22   0   
#>  9 1     -55.0 -2.81 2017-12-23   0   
#> 10 1     -55.0 -2.81 2017-12-24   7.37
#> # ... with 41 more rows

# S3 method for objects of class 'sf'
library("sf")

#> Linking to GEOS 3.6.1, GDAL 2.2.3, PROJ 4.9.3

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))

dates <- c("2017-12-15","2017-12-31")

get_chirps(lonlat, dates)

#> Getting your request...

#> # A tibble: 51 x 5
#>    id      lon   lat date       chirps
#>    <chr> <dbl> <dbl> <date>      <dbl>
#>  1 1     -55.0 -2.81 2017-12-15   0   
#>  2 1     -55.0 -2.81 2017-12-16   0   
#>  3 1     -55.0 -2.81 2017-12-17  13.7 
#>  4 1     -55.0 -2.81 2017-12-18  13.7 
#>  5 1     -55.0 -2.81 2017-12-19   0   
#>  6 1     -55.0 -2.81 2017-12-20   0   
#>  7 1     -55.0 -2.81 2017-12-21   0   
#>  8 1     -55.0 -2.81 2017-12-22   0   
#>  9 1     -55.0 -2.81 2017-12-23   0   
#> 10 1     -55.0 -2.81 2017-12-24   7.37
#> # ... with 41 more rows

# as.sf TRUE returns an object of class 'sf'
get_chirps(lonlat, dates, as.sf = TRUE)

#> Getting your request...

#> Simple feature collection with 3 features and 17 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -55.0714 ymin: -3.5279 xmax: -54.9857 ymax: -2.8094
#> epsg (SRID):    NA
#> proj4string:    NA
#>   d_17515 d_17516  d_17517  d_17518 d_17519 d_17520 d_17521 d_17522 d_17523
#> 1       0       0 13.68630 13.68630       0       0       0       0       0
#> 2       0       0 29.59562 14.79781       0       0       0       0       0
#> 3       0       0 11.06336 22.12671       0       0       0       0       0
#>    d_17524  d_17525  d_17526  d_17527   d_17528 d_17529   d_17530   d_17531
#> 1 7.373443 0.000000 0.000000 37.68985 18.844923       0  9.422462  9.422462
#> 2 0.000000 0.000000 9.385599 37.54240 18.771198       0  9.385599  9.385599
#> 3 0.000000 6.649946 0.000000 22.16965  5.542412       0 11.084825 11.084825
#>                   geometry
#> 1 POINT (-55.0281 -2.8094)
#> 2 POINT (-54.9857 -2.8756)
#> 3 POINT (-55.0714 -3.5279)