Get CHIRPS precipitation data

Get daily precipitation data from the "Climate Hazards Group InfraRed Precipitation with Station Data" via ClimateSERV API client. ClimateSERV works with geojson of type 'Polygon'. The input object is then transformed into polygons with a small buffer area around the point.

get_chirps(object, dates, operation = 5, ...)

# S3 method for default
get_chirps(object, dates, operation = 5, ...)

# S3 method for sf
get_chirps(object, dates, operation = 5, as.sf = FALSE, ...)

# S3 method for geojson
get_chirps(object, dates, operation = 5, as.geojson = FALSE, ...)

Arguments

object
input, an object of class data.frame, geojson or sf

dates
a character of start and end dates in that order in the format "YYYY-MM-DD"

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

...
further arguments passed to sf methods. See details

as.sf
logical, returns an object of class [sf]{sf} for S3 method of chirps class sf

as.geojson
logical, returns an object of class geojson for S3 method of chirps class geojson

Value

A data frame of CHIRPS data:

id

the index for the rows in object

dates

the dates from which CHIRPS was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

chirps

the CHIRPS value in mm

Details

operation: supported operations are:

operation		value

max	=	0

min	=	1

median	=	2

sum	=	4

average	=	5 (default value)

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

References

Funk C. et al. (2015). Scientific Data, 2, 150066.
https://doi.org/10.1038/sdata.2015.66

ClimateSERV https://climateserv.servirglobal.net

Examples

# \donttest{ 
#' library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857),
                     lat = c(-2.8094, -2.8756))

dates <- c("2017-12-15", "2017-12-31")

dat <- get_chirps(lonlat, dates)

############################################

# S3 method for objects of class 'sf'
library("sf")

# geometry 'POINT'
lonlat <- data.frame(lon = c(-55.0281, -55.0714),
                     lat = c(-2.8094,  -3.5279))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))

dates <- c("2017-12-15", "2017-12-31")

dat <- get_chirps(lonlat, dates)

# as.sf = TRUE returns an object of class 'sf'
get_chirps(lonlat, dates, as.sf = TRUE)

# geometry 'POLYGON'
p1 <- matrix(c(10.67, 49.90,
               10.57, 49.80,
               10.47, 49.90,
               10.57, 50.00,
               10.67, 49.90),
             nrow = 5, ncol = 2, byrow = TRUE)

p1 <- st_polygon(list(p1))

p2 <- matrix(c(11.67, 45.59,
               11.57, 45.49,
               11.47, 45.59,
               11.57, 45.69,
               11.67, 45.59),
             nrow = 5, ncol = 2, byrow = TRUE)

p2 <- st_polygon(list(p2))

pol <- data.frame(x = c(1, 2))
pol$geometry = st_sfc(p1, p2)

pol <- st_as_sf(pol)

dat <- get_chirps(pol, dates = c("2018-01-01", "2018-01-20"))

############################################

# S3 method for objects of class 'geojson'
library("sf")

tapajos <- chirps:::tapajos

object <- chirps:::.sf_to_geojson(tapajos[1:2,])

object <- unlist(object)

class(object) <- c("geojson", "json", class(object))

dates <- c("2018-01-01","2018-01-20")

dat <- get_chirps(object, dates)

# as.geojson = TRUE returns an object of class 'geojson'
dat <- get_chirps(object, dates, as.geojson = TRUE)

# }