Get Evaporative Stress Index (ESI) data

Get Evaporative Stress Index (ESI) from SERVIR Global via ClimateSERV API Client. ESI is available every four (or twelve) weeks from 2001 to present. The dataset may contain cloudy data which is returned as NAs. ClimateSERV works with geojson of type 'Polygon'. The input object is then transformed into polygons with a small buffer area around the point.

get_esi(object, dates, operation = 5, period = 1, ...)

# S3 method for default
get_esi(object, dates, operation = 5, period = 1, ...)

# S3 method for sf
get_esi(object, dates, operation = 5, period = 1, as.sf = FALSE, ...)

Arguments

object
input, an object of class data.frame or sf

dates
a character of start and end dates in that order in the format YYYY-MM-DD

operation
optional, an integer that represents which type of statistical operation to perform on the dataset

period
an integer value for the period of ESI data, four weeks period = 1, twelve weeks = 2

...
further arguments passed to sf methods. See details

as.sf
logical, returns an object of class "sf" for S3 method of class sf

Value

A data frame of ESI data:

id

the index for the rows in object

dates

the dates from which ESI was requested

lon

the longitude as provided in object

lat

the latitude as provided in object

esi

the ESI value

Details

operation: supported operations are max = 0, min = 1, median = 2, sum = 4, average = 5

dist: numeric, buffer distance for each object coordinate

nQuadSegs: integer, number of segments per buffer quadrant

References

ClimateSERV <https://climateserv.servirglobal.net>

Evaporative Stress Index <http://catalogue.servirglobal.net/Product?product_id=198>

Examples

library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15","2018-06-20")

# by default the function set a very small buffer around the points
# which can return NAs due to cloudiness in ESI data

dat <- get_esi(lonlat, dates = dates)

#> Getting your request...

dat

#> # A tibble: 41 x 5
#>    id      lon   lat date           esi
#>    <chr> <dbl> <dbl> <date>       <dbl>
#>  1 1     -55.0 -2.81 NA         NA     
#>  2 2     -55.0 -2.88 2017-12-11 -0.844 
#>  3 2     -55.0 -2.88 2018-01-01 -1.51  
#>  4 2     -55.0 -2.88 2018-01-09 -0.536 
#>  5 2     -55.0 -2.88 2018-01-17 -0.935 
#>  6 2     -55.0 -2.88 2018-01-25 -0.838 
#>  7 2     -55.0 -2.88 2018-02-02 -0.0731
#>  8 2     -55.0 -2.88 2018-03-06  1.27  
#>  9 2     -55.0 -2.88 2018-03-14  1.10  
#> 10 2     -55.0 -2.88 2018-03-22  0.792 
#> # ... with 31 more rows

# the argument dist passed through sf increase the buffer area

dat <- get_esi(lonlat, dates = dates, dist = 0.1)

#> Getting your request...

dat

#> # A tibble: 69 x 5
#>    id      lon   lat date           esi
#>    <chr> <dbl> <dbl> <date>       <dbl>
#>  1 1     -55.0 -2.81 2017-12-11 -1.98  
#>  2 1     -55.0 -2.81 2018-01-01 -1.49  
#>  3 1     -55.0 -2.81 2018-01-09 -0.219 
#>  4 1     -55.0 -2.81 2018-01-17 -0.423 
#>  5 1     -55.0 -2.81 2018-01-25 -0.239 
#>  6 1     -55.0 -2.81 2018-02-02  0.418 
#>  7 1     -55.0 -2.81 2018-02-10  0.435 
#>  8 1     -55.0 -2.81 2018-02-18  0.473 
#>  9 1     -55.0 -2.81 2018-02-26  0.126 
#> 10 1     -55.0 -2.81 2018-03-06  0.0570
#> # ... with 59 more rows