Precipitation indices

Compute precipitation indices over a time series.

precip_indices(object, timeseries = FALSE, span = NULL)

Arguments

object
a data.frame with CHIRPS data as privided by get_chirps

timeseries
logical, FALSE for a single point time series observation or TRUE for a time series based on a span

span
integer no lower than 5, for the days intervals when timeseries = TRUE

Value

A dataframe with selected indices:

MLDS

maximum length of consecutive dry days (r < 1 mm)

MLWS

maximum length of consecutive wet days (r >= 1 mm)

R10mm

number of heavy precipitation days (10 >= r < 20 mm)

R20mm

number of very heavy precipitation days (r >= 20)

Rx1day

maximum 1-day rainfall (mm)

Rx5day

maximum 5-day rainfall (mm)

SDII

simple daily intensity index (mean of wet days / total rainfall)

Rtotal

total rainfall (mm) in wet days (R >= 1)

References

Aguilar E., et al. (2005). Journal of Geophysical Research, 110(D23), D23107. https://doi.org/10.1029/2005JD006119

Kehel Z., et al. (2016). In: Applied Mathematics and Omics to Assess Crop Genetic Resources for Climate Change Adaptive Traits (eds Bari A., Damania A. B., Mackay M., Dayanandan S.), pp. 151–174. CRC Press.

See also

Examples

# Three points in the Tapajos National Forest, Brazil
lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15","2018-05-31")

df <- get_chirps(lonlat, dates)

# take the indices for the entire period 
precip_indices(df, timeseries = FALSE)

#> # A tibble: 24 x 6
#>       id   bin   lon   lat index   value
#>    <int> <dbl> <dbl> <dbl> <chr>   <dbl>
#>  1     1     1 -55.0 -2.81 MLDS      6  
#>  2     1     1 -55.0 -2.81 MLWS     11  
#>  3     1     1 -55.0 -2.81 R10mm    25  
#>  4     1     1 -55.0 -2.81 R20mm    30  
#>  5     1     1 -55.0 -2.81 Rx1day   72.1
#>  6     1     1 -55.0 -2.81 Rx5day  174. 
#>  7     1     1 -55.0 -2.81 SDII     20.1
#>  8     1     1 -55.0 -2.81 Rtotal 1585. 
#>  9     2     1 -55.0 -2.88 MLDS      7  
#> 10     2     1 -55.0 -2.88 MLWS      7  
#> # ... with 14 more rows

# take the indices for periods of 7 days
precip_indices(df, timeseries = TRUE, span = 7)

#> # A tibble: 576 x 6
#>       id   bin   lon   lat index  value
#>    <int> <int> <dbl> <dbl> <chr>  <dbl>
#>  1     1     1 -55.0 -2.81 MLDS     3  
#>  2     1     1 -55.0 -2.81 MLWS     2  
#>  3     1     1 -55.0 -2.81 R10mm    2  
#>  4     1     1 -55.0 -2.81 R20mm    0  
#>  5     1     1 -55.0 -2.81 Rx1day  13.7
#>  6     1     1 -55.0 -2.81 Rx5day  27.4
#>  7     1     1 -55.0 -2.81 SDII    13.7
#>  8     1     1 -55.0 -2.81 Rtotal  27.4
#>  9     1     2 -55.0 -2.81 MLDS     2  
#> 10     1     2 -55.0 -2.81 MLWS     2  
#> # ... with 566 more rows